\name{layernumber}
\alias{layernumber}
\alias{partitions}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Functions Extending Package vegsoup
}
\description{
\code{layernumber} returns layer assignments as ordered integer.
\code{partition} returns the number of samples per cluster.
}
\usage{
layernumber(obj)
partitions(obj)
}

\arguments{
  \item{obj}{
\code{VegsoupPartition*} object.
}
}
\value{
\code{numeric} vector.
}
\author{
Roland Kaiser
}
\seealso{
\code{\link{layers}},
\code{\link{partitioning}}
}
\examples{
# Nothing here yet.
}
\keyword{ manip }
