\name{synoptic-package}
\alias{synoptic-package}
\docType{package}
\title{
\packageTitle{synoptic}
}
\description{
\packageDescription{synoptic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{synoptic}
\packageIndices{synoptic}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{synoptic}

Maintainer: \packageMaintainer{synoptic}
}
\keyword{ package }

\examples{
# Nothing here yet
}
