\name{synoptic}
\alias{synoptic}
\alias{print.synoptic}

\title{
Create synoptic objects.
}
\description{
Create an object of class \code{synoptic} from an \code{\linkS4class{VegsoupPartitionFidelity}} object.
}
\usage{
synoptic(obj, stat.min = 0.4, p.max = 0.05)
}
\arguments{
  \item{obj}{
\code{VegsoupFidelity} object.
}
  \item{stat.min}{
\code{numeric} threshold value of fidelity statistics.
}
  \item{p.max}{
\code{numeric} threshold value of fisher test.
}
}
\details{
Nothing here yet.
}
\value{
An object of class \code{synoptic}.
}
\author{
Roland Kaiser
}
\examples{
# Nothing here yet
}
\keyword{ methods }
